requirements.txt
absl-py==2.2.0
appnope==0.1.4
arviz==0.21.0
asttokens==2.0.5
attrs==25.3.0
beartype==0.20.2
beautifulsoup4==4.13.4
bleach==6.2.0
Brotli==1.0.9
chex==0.1.89
cloudpickle==3.1.1
cola-ml==0.0.7
cola-plum-dispatch==0.1.4
comm==0.2.2
contourpy==1.3.1
CUQIpy==1.3.0.
CUQIpy-PyTorch==0.4.0
cycler==0.11.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.9
etils==1.12.2
executing==0.8.3
fastjsonschema==2.21.1
filelock==3.18.0
flax==0.10.4
fonttools==4.55.3
fsspec==2025.3.0
gast==0.6.0
gpjax==0.11.0
h5netcdf==1.6.1
h5py==3.13.0
humanize==4.12.2
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipympl==0.9.7
ipython==8.30.0
ipywidgets==8.1.7
jax==0.5.3
jaxlib==0.5.3
jaxopt==0.8.3
jaxtyping==0.3.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
multipledispatch==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.4.2
numpy==2.2.5
numpyro==0.18.0
opt_einsum==3.4.0
optax==0.2.4
optree==0.14.1
orbax-checkpoint==0.11.10
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.8.0
pillow==11.1.0
pip==25.0
platformdirs==4.3.7
prompt-toolkit==3.0.43
protobuf==6.30.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.19.1
pyparsing==3.2.0
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
rich==13.9.4
rpds-py==0.24.0
scikit-image==0.25.2
scipy==1.15.2
setuptools==75.8.0
simplejson==3.20.1
six==1.16.0
soupsieve==2.7
stack-data==0.2.0
sympy==1.13.1
tensorflow-probability==0.25.0
tensorstore==0.1.72
tifffile==2025.5.10
tinycss2==1.4.0
toolz==1.0.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
treescope==0.1.9
typing_extensions==4.12.2
tzdata==2025.2
unicodedata2==15.1.0
wadler_lindig==0.1.4
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.2
xarray==2025.3.0
xarray-einstats==0.8.0
zipp==3.21.0
